requirements.txt
pandas
numpy
faker
fpdf
psycopg2-binary
streamlit
openai
python-dotenv